Modified subscribe() to validate topics before sending subscriptions to MQTT servers
