+++ test/abstract_client.js
@@ -259,15 +259,21 @@ module.exports = function (server, config) {
 
     it('should be ok for well-formated topics', function (done) {
       var client = connect();
-      client.subscribe( ['+', '+/event', 'event/+', '#', 'event/#'], function (err) {
-        if (err) {
-          return done(new Error(err));
+      client.subscribe(
+        [
+          '+', '+/event', 'event/+', '#', 'event/#', 'system/event/+',
+          'system/+/event', 'system/registry/event/#', 'system/+/event/#'
+        ],
+        function (err) {
+          if (err) {
+            return done(new Error(err));
+          }
+          done();
         }
-        done();
-      });
+      );
     });
 
-    it('otherwise, should return an error (via callbacks)', function (done) {
+    it('should return an error (via callbacks) for topic #/event', function (done) {
       var client = connect();
       client.subscribe( ['#/event', 'event#', 'event+'], function (err) {
         if (err) {
@@ -276,6 +282,49 @@ module.exports = function (server, config) {
         done(new Error('Validations do NOT work'));
       });
     });
+
+    it('should return an error (via callbacks) for topic #/event', function (done) {
+      var client = connect();
+      client.subscribe( '#/event', function (err) {
+        if (err) {
+          return done();
+        }
+        done(new Error('Validations do NOT work'));
+      });
+    });
+
+    it('should return an error (via callbacks) for topic event#', function (done) {
+      var client = connect();
+      client.subscribe( 'event#', function (err) {
+        if (err) {
+          return done();
+        }
+        done(new Error('Validations do NOT work'));
+      });
+    });
+
+    it('should return an error (via callbacks) for topic system/#/event', function (done) {
+      var client = connect();
+      client.subscribe( 'system/#/event', function (err) {
+        if (err) {
+          return done();
+        }
+        done(new Error('Validations do NOT work'));
+      });
+    });
+
+
+    it('should return an error (via callbacks) for topic system/+/#/event', function (done) {
+      var client = connect();
+      client.subscribe( 'system/+/#/event', function (err) {
+        if (err) {
+          return done();
+        }
+        done(new Error('Validations do NOT work'));
+      });
+    });
+
+
 
   });
